+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.8.7] - 2026-07-10
+
+### Added
+
+- Pages can now set their own response headers, like how long browsers should cache them.
+
+---
+
 ## [0.8.6] - 2026-07-05
 
 ### Added